requirements.txt
jsonschema
pytest
abacusai
fastapi
uvicorn
pdfplumber
python-docx
pandas
requests

